% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2Matrix.R
\name{List2Matrix}
\alias{List2Matrix}
\title{Convert a list of matrices into a single large matrix}
\usage{
List2Matrix(x, sym = FALSE)
}
\arguments{
\item{x}{A numeric list with each entity be a numeric matrix with p rows and q columns}

\item{sym}{A logical scalar representing whether each matrix is symmetric. If true, the duplicated half is removed}
}
\value{
A numeric matrix with pq rows and T columns
}
\description{
Convert a list of matrices into a single large matrix
}
\examples{
x=list(x1=1:3,x2=4:6)
List2Matrix(x)

y=list(y1=matrix(1:4,2),y2=matrix(5:8,2))
List2Matrix(y)

}
